backend/requirements.txt
asgiref==3.8.1
attrs==25.3.0
black==25.1.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
defusedxml==0.7.1
Django==4.2
django-cors-headers==4.3.1
django-filter==24.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-extra-fields==3.7.0
drf-spectacular==0.28.0
filetype==1.2.0
gunicorn==22.0.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
mypy_extensions==1.1.0
oauthlib==3.2.2
packaging==25.0
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.3.7
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.10.1
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.24.0
social-auth-app-django==5.4.3
social-auth-core==4.6.0
sqlparse==0.5.3
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0